fix some jpms issues for websocket-core
