+++ jetty-core/jetty-websocket/jetty-websocket-core-server/src/main/java/org/eclipse/jetty/websocket/core/server/internal/WebSocketNegotiation.java
@@ -45,8 +45,8 @@ public abstract class WebSocketNegotiation
 
     public WebSocketNegotiation(Request request, Response response, Callback callback, WebSocketComponents webSocketComponents)
     {
-        this.request = new ServerUpgradeRequest(this, request);
-        this.response = new ServerUpgradeResponse(this, response);
+        this.request = new ServerUpgradeRequestImpl(this, request);
+        this.response = new ServerUpgradeResponseImpl(this, response);
         this.callback = callback;
         this.components = webSocketComponents;
     }
